Overhaul settings, add support for admins and super admins

+++ index.js
@@ -15,8 +15,14 @@ try {
         settings = fs.readFileSync('settings.json');
         settings = JSON.parse(settings);
     } else {
-        settings = makeSettings();
-        fs.writeFileSync(settingsPath, JSON.stringify(settings, null, 4));
+        if(fs.existsSync('settingsBase.json')){
+            settings = fs.readFileSync('settingsBase.json');
+            settings = JSON.parse(settings);
+            fs.writeFileSync(settingsPath, JSON.stringify(settings, null, 4));
+        }
+        else{
+            console.log("Error, could not find base settings!");
+        }
     }
 } catch(err){
     console.log(err);
@@ -26,13 +32,6 @@ const webhook = require("webhook-discord");
 
 // Create a client with our options
 const client = new tmi.client(settings.connectionSettings);
-const _personalAliases = ["durok"];
-const _mcdmAliases = ["matt","mcdm"];
-const personalAliases = _personalAliases.join("|");
-const mcdmAliases = _mcdmAliases.join("|");
-
-let questionsOpen = false;
-let listenForMentions = true;
 let date = new Date();
 let cooldown = 5000;
 let lastCommandTime = 0;
@@ -57,47 +56,99 @@ function onMessageHandler (target, context, msg, self) {
     const commandTriggered = messageString.split(" ")[0].toLowerCase();
     //get the settings for the channel
     //let commandSettingsName = settings.commands.find((command) => command.name == commandTriggered);
-    let command = settings.commands.find((command) => command.channels.find((channel) => channel == target.substr(1)) && command.command == commandTriggered);
-    if(command === undefined)
-    {
-        console.log("command settings not found");
+    let command = settings.commands[commandTriggered];
+    if(command === undefined){
+        console.log("command not found");
         return;
     }
-        
-    if(command.settings.requiresMod === true){
-        if(checkIfMod(context, target) === false){
-            client.say(target, "@"+context["display-name"]+" you lack the correct Choob permissions to execute this command!");
+    if(command.choobbotChannelOnly === true){
+        if(checkIfChoobbotChannel(target) === false){
+            //client.say(target, "@"+context["display-name"]+" you lack the correct Choob permissions to execute this command!");
             return;
         }
     }
-    // If the command is known, let's execute it
-    switch (commandTriggered){
+    let isUserSuperAdmin = checkIfSuperAdmin(context);
+    
+    if(!isUserSuperAdmin){
+        if(command.requiresSuperAdmin === true){
+            client.say(target, "@"+context["display-name"]+" you lack the correct Choobs to execute this command!");
+            return;
+        }
+        if(command.requiresAdmin === true){
+            if(checkIfAdmin(context, target) === false){
+                client.say(target, "@"+context["display-name"]+" you lack the correct Choobs to execute this command!");
+                return;
+            }
+        }
+        if(command.requiresMod === true){
+            if(checkIfMod(context, target) === false){
+                client.say(target, "@"+context["display-name"]+" you lack the correct Choobs to execute this command!");
+                return;
+            }
+        }
+    }
+    
+    switch(commandTriggered){
         case "choob":
+            if(command.ignoredChannels.includes(target))
+                break;
             if(lastCommandTime + cooldown < date.getTime()){
-                let i = command.settings.messages.length;
-                let s = command.settings.messages[Math.floor(Math.random() * i)];
+                let i = command.messages.length;
+                let s = command.messages[Math.floor(Math.random() * i)];
                 client.say(target, s);
                 console.log(`* Executed ${commandTriggered} command`);
             }
             break;
+        case "joinchoob": {
+            let channelstring = "#" + context.username;
+            if(settings.connectionSettings.channels.includes(channelstring)){
+                client.say(target, "Choob Bot is already in your channel!");
+            } else {
+                settings.connectionSettings.channels.push(channelstring);
+                fs.writeFile(settingsPath, JSON.stringify(settings, null, 4), (e) => {if(e != null)console.log(e);});
+                client.say(target, "Choob Bot has joined your channel!");
+                client.join(channelstring);
+            }
+            console.log(`* Executed ${commandTriggered} command`);
+            break;
+        }
+        case "leavechoob": {
+            const removalChannel = "#" + context.username;
+            let i = settings.connectionSettings.channels.indexOf(removalChannel);
+            if(i != -1){
+                settings.connectionSettings.channels.splice(i, 1);
+                fs.writeFile(settingsPath, JSON.stringify(settings, null, 4), (e) => {if(e != null)console.log(e);});
+                client.say(target, "Choob Bot has left your channel...");
+                client.part(removalChannel);
+                console.log(`* Executed ${commandTriggered} command`);
+            }
+            break;
+        }
         case "choobcount":
             if(lastCommandTime + cooldown < date.getTime()){
-                let choobCommand = settings.commands.find((command) => command.command == "choob");
-                let i = choobCommand.settings.messages.length;
+                let choobCommand = settings.commands.choob;
+                let i = choobCommand.messages.length;
                 client.say(target, "I have " + i + " Choobs!");
                 console.log(`* Executed ${commandTriggered} command`);
             }
             break;
+        case "choobinfo":
+            if(lastCommandTime + cooldown < date.getTime()){
+                let info = command.settings.message;
+                client.say(target, info);
+                console.log(`* Executed ${commandTriggered} command`);
+            }
+            break;
         case "togglechoob":
-            let choobCommand = settings.commands.find((command) => command.command == "choob");
-            let exists = choobCommand.channels.indexOf(target.substr(1));
+            let choobCommand = settings.commands.choob;
+            let exists = choobCommand.ignoredChannels.indexOf(target);
             if(exists >= 0){
-                choobCommand.channels.splice(exists, 1);
-                client.say(target, "!Choob has been disabled");
+                choobCommand.ignoredChannels.splice(exists, 1);
+                client.say(target, "!Choob has been enabled");
                 
             } else {
-                choobCommand.channels.push(target.substr(1));
-                client.say(target, "!Choob has been enabled");
+                choobCommand.ignoredChannels.push(target);
+                client.say(target, "!Choob has been disabled");
             }
             fs.writeFile(settingsPath, JSON.stringify(settings, null, 4), (e) => {if(e != null)console.log(e);});
             console.log(`* Executed ${commandTriggered} command`);
@@ -107,9 +158,8 @@ function onMessageHandler (target, context, msg, self) {
                 break;
             }
             const choobString = messageString.substr(commandTriggered.length + 1); //remove commnad name and first space
-
-            let choobc = settings.commands.find((command) => command.command == "choob");
-            choobc.settings.messages.push(choobString);
+            let choobc = settings.commands.choob;
+            choobc.messages.push(choobString);
             fs.writeFile(settingsPath, JSON.stringify(settings, null, 4), (e) => {if(e != null)console.log(e);});
             client.say(target, "Added Choob: \""+choobString+"\" to the master Choob list!");
             console.log(`* Executed ${commandTriggered} command`);
@@ -119,17 +169,17 @@ function onMessageHandler (target, context, msg, self) {
                 break;
             }
             const removalString = messageString.substr(commandTriggered.length + 1); //remove commnad name and first space
-            let choobCommandInfo = settings.commands.find((command) => command.command == "choob");
+            let choobCommandInfo = settings.commands.choob;
             
-            let i = choobCommandInfo.settings.messages.indexOf(removalString);
+            let i = choobCommandInfo.messages.indexOf(removalString);
             if(i != -1){
-                choobCommandInfo.settings.messages.splice(i, 1);
+                choobCommandInfo.messages.splice(i, 1);
                 fs.writeFile(settingsPath, JSON.stringify(settings, null, 4), (e) => {if(e != null)console.log(e);});
                 client.say(target, "Removed Choob: \""+removalString+"\" from the master Choob list!");
                 console.log(`* Executed ${commandTriggered} command`);
             }
             break;
-        case "updatesettings":
+        case "updatechoob":
             try {
                 settings = fs.readFileSync('settings.json');
                 settings = JSON.parse(settings);
@@ -144,6 +194,7 @@ function onMessageHandler (target, context, msg, self) {
         default:
             console.log(`* Unknown command ${commandTriggered}`);
     }
+    return;
 }
 
 // Called every time the bot connects to Twitch chat
@@ -153,28 +204,14 @@ function onConnectedHandler (addr, port) {
 function checkIfMod(context, target){
     return context.mod == true || context.username == target.substr(1);
 }
-
-function makeChoobSettings(){
-    let choobSettings = {
-        currentlyListening : true,
-        messages : ["Choob!"]
-    };
-    return choobSettings;
+function checkIfAdmin(context, target){
+    if(checkIfMod(context, target) === true && settings.adminChannels.includes(target))
+        return true;
+    return false;
 }
-
-function makeCommands(){
-    let commands = {
-        channel : "",
-        choobSettings : null,
-        ignoredUsers : []
-    };
-    commands.choobSettings = makeChoobSettings();
-    return commands;
+function checkIfSuperAdmin(context){
+    return settings.superAdmins.includes(context.username);
 }
-function makeSettings(){
-    let settings = {
-        commands : []
-    };
-    settings.commands[0] = makeCommands();
-    return settings;
+function checkIfChoobbotChannel(target){
+    return target === "#choob_bot";
 }
